requirements.txt
# requirements.txt

# Python packages
python-pptx==0.6.19
Pillow==9.0.0
selenium==4.1.0
chromedriver-autoinstaller==0.0.6